requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
async-generator==1.10
async-lru==2.0.2
async-timeout==4.0.2
attrs==22.2.0
azure-core==1.26.4
azure-storage-blob==12.16.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
black==23.3.0
bleach==4.1.0
blinker==1.8.2
browser-cookie3==0.19.1
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
cleanco==2.2
cliapp==1.20180812.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
command-not-found==0.3
commonmark==0.9.1
contourpy==1.0.7
cookiecutter==2.1.1
cryptography==3.4.8
cw-etl-utils==1.0.0
cycler==0.11.0
databricks-cli==0.17.6
dbus-python==1.2.18
DBUtils==3.0.3
dbx==0.8.10
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
delta-spark==2.4.0
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.2
findspark==2.0.1
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gpg==1.16.0
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
identify==2.5.24
idna==3.4
importlib-metadata==4.6.4
install==1.3.5
ipykernel==6.22.0
ipython==8.12.0
ipython_genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
jeepney==0.7.1
jellyfish==0.11.2
Jinja2==3.1.4
jinja2-time==0.2.0
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.23.0
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libcomps==0.1.15
linkedin_api==2.1.1
lxml==4.9.2
lz4==4.3.3
Markdown==3.3.6
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mlflow-skinny==2.3.0
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.5.6
nbconvert==7.5.0
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.11.0
nodeenv==1.8.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.2
oauthlib==3.2.0
openai==1.31.2
openpyxl==3.1.2
outcome==1.2.0
overrides==7.3.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.6.0
plotly==5.14.1
pre-commit==3.3.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
py4j==0.10.9.7
pyarrow==11.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.20.0
pydantic==1.10.7
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.19.3
PySelenium==1.0.2
PySocks==1.7.1
pyspark==3.4.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
rapidfuzz==3.5.2
recordlinkage==0.15
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rpm==4.17.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==3.3.1
selenium==4.10.0
Send2Trash==1.8.2
shellingham==1.5.0.post1
simple-salesforce==1.12.5
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
stack-data==0.6.2
systemd-python==234
tabulate==0.9.0
tenacity==8.2.2
terminado==0.17.1
testpath==0.5.0
text-unidecode==1.3
thefuzz==0.20.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenize-rt==5.0.0
tomli==2.0.1
tornado==6.2
tqdm==4.66.4
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
ttystatus==0.38
typer==0.7.0
typing_extensions==4.10.0
tzdata==2023.3
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.23.1
wadllib==1.3.6
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.3
Werkzeug==3.0.3
widgetsnbextension==4.0.7
wsproto==1.2.0
yarl==1.8.2
zeep==4.2.1
zipp==1.0.0